requirements.txt
python-telegram-bot==20.7
google-generativeai==0.6.0